ai-service/requirements.txt
# ai-service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
ollama
openai-whisper 
pydub 
ffmpeg-python
python-multipart